Move TM2 log to common
Fixes #1919
